+++ src/lib.rs
@@ -1,40 +1,10 @@
 pub mod rustdoc_json_types;
+pub mod render_core;
+pub mod render_details;
+pub mod render_signatures;
+
 use rustdoc_json_types::*;
-
-// Helper struct to carry context about how an item is being presented
-#[derive(Clone, Debug)]
-struct ResolvedItemInfo<'a> {
-    original_item: &'a Item,
-    effective_name: Option<String>, // Name in the current scope (alias or original)
-    reexport_source_canonical_path: Option<String>, // Canonical path of the original item, if re-exported
-}
-
-fn get_item_kind_string(item_enum: &ItemEnum) -> &str {
-    match item_enum {
-        ItemEnum::Module(_) => "Module",
-        ItemEnum::Struct(_) => "Struct",
-        ItemEnum::Enum(_) => "Enum",
-        ItemEnum::Union(_) => "Union",
-        ItemEnum::Trait(_) => "Trait",
-        ItemEnum::TraitAlias(_) => "Trait Alias",
-        ItemEnum::Function(_) => "Function",
-        ItemEnum::TypeAlias(_) => "Type Alias",
-        ItemEnum::Constant { .. } => "Constant",
-        ItemEnum::Static(_) => "Static",
-        ItemEnum::Macro(_) => "Macro",
-        ItemEnum::ProcMacro(_) => "Procedural Macro",
-        ItemEnum::ExternCrate { .. } => "Extern Crate",
-        ItemEnum::Use(_) => "Use Statement", // Should primarily be handled by format_use_statement_for_listing
-        ItemEnum::StructField(_) => "Struct Field",
-        ItemEnum::Variant(_) => "Variant",
-        ItemEnum::Impl(_) => "Implementation",
-        ItemEnum::Primitive(_) => "Primitive",
-        ItemEnum::ExternType => "Extern Type",
-        ItemEnum::AssocConst { .. } => "Associated Constant",
-        ItemEnum::AssocType { .. } => "Associated Type",
-    }
-}
-
+use render_core::process_items;
 
 pub fn rustdoc_json_to_markdown(data: Crate) -> String {
     let mut output = String::new();
@@ -70,2276 +40,4 @@ pub fn rustdoc_json_to_markdown(data: Crate) -> String {
     }
 
     output
-}
-
-fn process_items(output: &mut String, item_ids: &[Id], data: &Crate, level: usize) {
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-
-    let mut all_resolved_items: Vec<ResolvedItemInfo<'_>> = Vec::new();
-    let mut use_item_ids_for_reexport_section: Vec<Id> = Vec::new();
-
-    for &id in item_ids {
-        if let Some(item) = data.index.get(&id) {
-            match &item.inner {
-                ItemEnum::Use(use_item) => {
-                    use_item_ids_for_reexport_section.push(id); // Always add Use item for "Re-exports" section
-
-                    // Phase 1: Process named re-exports for inlining
-                    if !use_item.is_glob {
-                        if let Some(target_id) = use_item.id {
-                            if let Some(target_item) = data.index.get(&target_id) {
-                                // Check if target is struct or enum as per Phase 1 focus,
-                                // but generally, we can try to inline any named re-export.
-                                // For Phase 1, let's stick to the plan's focus on structs/enums.
-                                match &target_item.inner {
-                                    ItemEnum::Struct(_) | ItemEnum::Enum(_) => {
-                                        let canonical_path = data.paths.get(&target_item.id)
-                                            .map(|summary| summary.path.join("::"))
-                                            .unwrap_or_else(|| target_item.name.clone().unwrap_or_default());
-
-                                        all_resolved_items.push(ResolvedItemInfo {
-                                            original_item: target_item,
-                                            effective_name: Some(use_item.name.clone()),
-                                            reexport_source_canonical_path: Some(canonical_path),
-                                        });
-                                    }
-                                    _ => {
-                                        // Other re-exported types will just be listed in the "Re-exports" section for now.
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    // Glob re-exports are just listed in the "Re-exports" section for Phase 1.
-                }
-                _ => {
-                    // Direct items
-                    all_resolved_items.push(ResolvedItemInfo {
-                        original_item: item,
-                        effective_name: item.name.clone(), // Use its own name
-                        reexport_source_canonical_path: None,
-                    });
-                }
-            }
-        }
-    }
-
-    // Group resolved items by kind for better organization
-    let mut modules: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-    let mut types: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-    let mut traits: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-    let mut functions: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-    let mut constants: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-    let mut macros: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-    // Note: 'reexports' (use_item_ids_for_reexport_section) are handled separately.
-    let mut other_items: Vec<&ResolvedItemInfo<'_>> = Vec::new();
-
-
-    for resolved_info in &all_resolved_items {
-        match &resolved_info.original_item.inner {
-            ItemEnum::Module(_) => modules.push(resolved_info),
-            ItemEnum::Struct(_) | ItemEnum::Enum(_) | ItemEnum::Union(_) | ItemEnum::TypeAlias(_) => {
-                types.push(resolved_info)
-            }
-            ItemEnum::Trait(_) | ItemEnum::TraitAlias(_) => traits.push(resolved_info),
-            ItemEnum::Function(_) => functions.push(resolved_info),
-            ItemEnum::Constant { .. } | ItemEnum::Static(_) => constants.push(resolved_info),
-            ItemEnum::Macro(_) | ItemEnum::ProcMacro(_) => macros.push(resolved_info),
-            ItemEnum::Use(_) => {} // These are handled by use_item_ids_for_reexport_section
-            _ => other_items.push(resolved_info),
-        }
-    }
-
-    // Process each group in order
-    if !modules.is_empty() {
-        output.push_str(&format!("{} Modules\n\n", "#".repeat(heading_level)));
-        for resolved_info in modules {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-
-    if !types.is_empty() {
-        output.push_str(&format!("{} Types\n\n", "#".repeat(heading_level)));
-        for resolved_info in types {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-
-    if !traits.is_empty() {
-        output.push_str(&format!("{} Traits\n\n", "#".repeat(heading_level)));
-        for resolved_info in traits {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-
-    if !functions.is_empty() {
-        output.push_str(&format!("{} Functions\n\n", "#".repeat(heading_level)));
-        for resolved_info in functions {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-
-    if !constants.is_empty() {
-        output.push_str(&format!(
-            "{} Constants and Statics\n\n",
-            "#".repeat(heading_level)
-        ));
-        for resolved_info in constants {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-
-    if !macros.is_empty() {
-        output.push_str(&format!("{} Macros\n\n", "#".repeat(heading_level)));
-        for resolved_info in macros {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-
-    // Render the "Re-exports" section for transparency, listing the use statements themselves
-    if !use_item_ids_for_reexport_section.is_empty() {
-        output.push_str(&format!("{} Re-exports\n\n", "#".repeat(heading_level)));
-        for id in use_item_ids_for_reexport_section {
-            if let Some(use_item_obj) = data.index.get(&id) {
-                 format_use_statement_for_listing(output, use_item_obj, data, level + 1);
-            }
-        }
-    }
-
-    if !other_items.is_empty() {
-        output.push_str(&format!("{} Other Items\n\n", "#".repeat(heading_level)));
-        for resolved_info in other_items {
-            process_item(output, resolved_info, data, level + 1);
-        }
-    }
-}
-
-// New function to format just the 'use' statement for the "Re-exports" section
-fn format_use_statement_for_listing(output: &mut String, item: &Item, _data: &Crate, level: usize) {
-    let heading_level = std::cmp::min(level, 6); // Use level+1 from process_items for consistency
-    let heading = "#".repeat(heading_level);
-
-    if let ItemEnum::Use(use_item) = &item.inner {
-        let source_name_segment = use_item.source.split("::").last().unwrap_or(&use_item.source);
-
-        // Heading for the use statement
-        if use_item.is_glob {
-            output.push_str(&format!("{} `use {}::*`\n\n", heading, use_item.source));
-        } else {
-            // Use item.name for the alias, use_item.name is the same if not aliased
-            let display_name = item.name.as_ref().unwrap_or(&use_item.name);
-            if display_name != source_name_segment && item.name.is_some() { // Aliased
-                 output.push_str(&format!("{} `use {} as {}`\n\n", heading, use_item.source, display_name));
-            } else { // Not aliased or item.name is not the alias part
-                 output.push_str(&format!("{} `use {}`\n\n", heading, use_item.source));
-            }
-        }
-
-        // Add documentation for the use statement itself, if any
-        if let Some(docs) = &item.docs {
-            output.push_str(&format!("{}\n\n", docs));
-        }
-
-        // Add the 'use ...;' line in a code block
-        output.push_str("```rust\n");
-        let mut use_signature = String::new();
-        match &item.visibility {
-            Visibility::Public => use_signature.push_str("pub "),
-            Visibility::Crate => use_signature.push_str("pub(crate) "),
-            Visibility::Restricted { path, .. } => use_signature.push_str(&format!("pub(in {}) ", path)),
-            Visibility::Default => {}
-        }
-        use_signature.push_str(&format!("use {}", use_item.source));
-        if use_item.is_glob {
-            use_signature.push_str("::*");
-        } else if let Some(name_attr) = &item.name { // item.name here is the alias if `as Alias` was used
-            // Check if it's a renaming alias
-            if name_attr != source_name_segment {
-                 use_signature.push_str(&format!(" as {}", name_attr));
-            }
-        }
-        use_signature.push(';');
-        output.push_str(&use_signature);
-        output.push_str("\n```\n\n");
-    }
-}
-
-
-fn process_item(output: &mut String, resolved_info: &ResolvedItemInfo, data: &Crate, level: usize) {
-    let item = resolved_info.original_item;
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-    let heading_prefix = "#".repeat(heading_level);
-
-    let display_name_opt = resolved_info.effective_name.as_ref().or(item.name.as_ref());
-    let item_kind_str = get_item_kind_string(&item.inner);
-
-    // Add item heading with name and kind
-    if let Some(display_name) = display_name_opt {
-        if let Some(canonical_path) = &resolved_info.reexport_source_canonical_path {
-            // This is a re-exported item.
-            // Format: "Re-exported Kind `DisplayName` (from `CanonicalPath`)"
-            // If DisplayName is an alias (i.e., different from original item name),
-            // it's already captured in display_name.
-            output.push_str(&format!(
-                "{} Re-exported {} `{}` (from `{}`)\n\n",
-                heading_prefix, item_kind_str, display_name, canonical_path
-            ));
-        } else {
-            // This is a direct definition.
-            // Special handling for Module level to ensure it's ##
-            if let ItemEnum::Module(_) = &item.inner {
-                 output.push_str(&format!("## Module `{}`\n\n", display_name));
-            } else {
-                 output.push_str(&format!("{} {} `{}`\n\n", heading_prefix, item_kind_str, display_name));
-            }
-        }
-    } else {
-        // Handle nameless items like Impl
-        match &item.inner {
-            ItemEnum::Impl(impl_details) => {
-                if let Some(trait_) = &impl_details.trait_ {
-                    output.push_str(&format!(
-                        "{} Implementation of `{}` for `{}`\n\n",
-                        heading_prefix,
-                        trait_.path,
-                        format_type(&impl_details.for_, data)
-                    ));
-                } else {
-                    output.push_str(&format!(
-                        "{} Implementation for `{}`\n\n",
-                        heading_prefix,
-                        format_type(&impl_details.for_, data)
-                    ));
-                }
-            }
-            _ => {
-                // For other truly unnamed items, or items where name is not applicable for heading
-                output.push_str(&format!("{} {}\n\n", heading_prefix, item_kind_str));
-            }
-        }
-    }
-
-
-    // Add item attributes if present
-    if !item.attrs.is_empty() {
-        output.push_str("**Attributes:**\n\n");
-        for attr in &item.attrs {
-            output.push_str(&format!("- `{}`\n", attr));
-        }
-        output.push('\n');
-    }
-
-    // Add deprecation info if present
-    if let Some(deprecation) = &item.deprecation {
-        output.push_str("**⚠️ Deprecated");
-        if let Some(since) = &deprecation.since {
-            output.push_str(&format!(" since {}", since));
-        }
-        output.push_str("**");
-
-        if let Some(note) = &deprecation.note {
-            output.push_str(&format!(": {}", note));
-        }
-        output.push_str("\n\n");
-    }
-
-    // Add documentation if available
-    if let Some(docs) = &item.docs {
-        output.push_str(&format!("{}\n\n", docs));
-    }
-
-    // Add code block with item signature (using the original item for signature)
-    output.push_str("```rust\n");
-    format_item_signature(output, item, data); // Pass original 'item'
-    output.push_str("\n```\n\n");
-
-    // Process additional details based on item kind (using original item)
-    // Phase 1 focuses on basic rendering; details functions are not changed yet.
-    match &item.inner {
-        ItemEnum::Module(module) => process_module_details(output, module, data, level + 1),
-        ItemEnum::Struct(struct_) => process_struct_details(output, struct_, data, level + 1),
-        ItemEnum::Enum(enum_) => process_enum_details(output, enum_, data, level + 1),
-        ItemEnum::Union(union_) => process_union_details(output, union_, data, level + 1),
-        ItemEnum::Trait(trait_) => process_trait_details(output, trait_, data, level + 1),
-        ItemEnum::Impl(impl_) => process_impl_details(output, impl_, data, level + 1),
-        _ => {}
-    }
-}
-
-fn format_item_signature(output: &mut String, item: &Item, data: &Crate) {
-    // Format visibility
-    match &item.visibility {
-        Visibility::Public => output.push_str("pub "),
-        Visibility::Crate => output.push_str("pub(crate) "),
-        Visibility::Restricted { path, .. } => output.push_str(&format!("pub(in {}) ", path)),
-        Visibility::Default => {}
-    }
-
-    // Format item based on its kind
-    match &item.inner {
-        ItemEnum::Module(_) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("mod {} {{ /* ... */ }}", name));
-            }
-        }
-        ItemEnum::Struct(struct_) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("struct {}", name));
-                format_generics(output, &struct_.generics, data);
-
-                match &struct_.kind {
-                    StructKind::Unit => output.push(';'),
-                    StructKind::Tuple(fields) => {
-                        output.push('(');
-                        for (i, field_opt) in fields.iter().enumerate() {
-                            if let Some(field_id) = field_opt {
-                                if let Some(field_item) = data.index.get(field_id) {
-                                    if let ItemEnum::StructField(field_type) = &field_item.inner {
-                                        // Field visibility if needed
-                                        match &field_item.visibility {
-                                            Visibility::Public => output.push_str("pub "),
-                                            Visibility::Crate => output.push_str("pub(crate) "),
-                                            Visibility::Restricted { path, .. } => {
-                                                output.push_str(&format!("pub(in {}) ", path))
-                                            }
-                                            Visibility::Default => {}
-                                        }
-                                        output.push_str(&format_type(field_type, data));
-                                    }
-                                }
-                                if i < fields.len() - 1 {
-                                    output.push_str(", ");
-                                }
-                            } else {
-                                // For stripped fields
-                                output.push_str("/* private field */");
-                                if i < fields.len() - 1 {
-                                    output.push_str(", ");
-                                }
-                            }
-                        }
-                        output.push_str(");");
-                    }
-                    StructKind::Plain {
-                        fields,
-                        has_stripped_fields,
-                    } => {
-                        output.push_str(" {\n");
-                        for &field_id in fields {
-                            if let Some(field_item) = data.index.get(&field_id) {
-                                if let Some(field_name) = &field_item.name {
-                                    if let ItemEnum::StructField(field_type) = &field_item.inner {
-                                        // Field visibility
-                                        match &field_item.visibility {
-                                            Visibility::Public => output.push_str("    pub "),
-                                            Visibility::Crate => output.push_str("    pub(crate) "),
-                                            Visibility::Restricted { path, .. } => {
-                                                output.push_str(&format!("    pub(in {}) ", path))
-                                            }
-                                            Visibility::Default => output.push_str("    "),
-                                        }
-                                        output.push_str(&format!(
-                                            "{}: {},\n",
-                                            field_name,
-                                            format_type(field_type, data)
-                                        ));
-                                    }
-                                }
-                            }
-                        }
-                        if *has_stripped_fields {
-                            output.push_str("    // Some fields omitted\n");
-                        }
-                        output.push('}');
-                    }
-                }
-            }
-        }
-        ItemEnum::Enum(enum_) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("enum {}", name));
-                format_generics(output, &enum_.generics, data);
-                output.push_str(" {\n");
-
-                for &variant_id in &enum_.variants {
-                    if let Some(variant_item) = data.index.get(&variant_id) {
-                        if let Some(variant_name) = &variant_item.name {
-                            output.push_str(&format!("    {}", variant_name));
-
-                            if let ItemEnum::Variant(variant) = &variant_item.inner {
-                                match &variant.kind {
-                                    VariantKind::Plain => {}
-                                    VariantKind::Tuple(fields) => {
-                                        output.push('(');
-                                        for (i, field_opt) in fields.iter().enumerate() {
-                                            if let Some(field_id) = field_opt {
-                                                if let Some(field_item) = data.index.get(field_id) {
-                                                    if let ItemEnum::StructField(field_type) =
-                                                        &field_item.inner
-                                                    {
-                                                        output.push_str(&format_type(
-                                                            field_type, data,
-                                                        ));
-                                                    }
-                                                }
-                                                if i < fields.len() - 1 {
-                                                    output.push_str(", ");
-                                                }
-                                            } else {
-                                                // For stripped fields
-                                                output.push_str("/* private field */");
-                                                if i < fields.len() - 1 {
-                                                    output.push_str(", ");
-                                                }
-                                            }
-                                        }
-                                        output.push(')');
-                                    }
-                                    VariantKind::Struct {
-                                        fields,
-                                        has_stripped_fields,
-                                    } => {
-                                        output.push_str(" {\n");
-                                        for &field_id in fields {
-                                            if let Some(field_item) = data.index.get(&field_id) {
-                                                if let Some(field_name) = &field_item.name {
-                                                    if let ItemEnum::StructField(field_type) =
-                                                        &field_item.inner
-                                                    {
-                                                        output.push_str(&format!(
-                                                            "        {}: {},\n",
-                                                            field_name,
-                                                            format_type(field_type, data)
-                                                        ));
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        if *has_stripped_fields {
-                                            output.push_str("        // Some fields omitted\n");
-                                        }
-                                        output.push_str("    }");
-                                    }
-                                }
-
-                                if let Some(discriminant) = &variant.discriminant {
-                                    output.push_str(&format!(" = {}", discriminant.expr));
-                                }
-                            }
-
-                            output.push_str(",\n");
-                        }
-                    }
-                }
-
-                if enum_.has_stripped_variants {
-                    output.push_str("    // Some variants omitted\n");
-                }
-
-                output.push('}');
-            }
-        }
-        ItemEnum::Union(union_) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("union {}", name));
-                format_generics(output, &union_.generics, data);
-                output.push_str(" {\n");
-
-                for &field_id in &union_.fields {
-                    if let Some(field_item) = data.index.get(&field_id) {
-                        if let Some(field_name) = &field_item.name {
-                            if let ItemEnum::StructField(field_type) = &field_item.inner {
-                                match &field_item.visibility {
-                                    Visibility::Public => output.push_str("    pub "),
-                                    Visibility::Crate => output.push_str("    pub(crate) "),
-                                    Visibility::Restricted { path, .. } => {
-                                        output.push_str(&format!("    pub(in {}) ", path))
-                                    }
-                                    Visibility::Default => output.push_str("    "),
-                                }
-                                output.push_str(&format!(
-                                    "{}: {},\n",
-                                    field_name,
-                                    format_type(field_type, data)
-                                ));
-                            }
-                        }
-                    }
-                }
-
-                if union_.has_stripped_fields {
-                    output.push_str("    // Some fields omitted\n");
-                }
-
-                output.push('}');
-            }
-        }
-        ItemEnum::Function(function) => {
-            // Function header
-            if function.header.is_const {
-                output.push_str("const ");
-            }
-            if function.header.is_unsafe {
-                output.push_str("unsafe ");
-            }
-            if function.header.is_async {
-                output.push_str("async ");
-            }
-
-            // ABI
-            match &function.header.abi {
-                Abi::Rust => {}
-                Abi::C { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"C-unwind\" ");
-                    } else {
-                        output.push_str("extern \"C\" ");
-                    }
-                }
-                Abi::Cdecl { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"cdecl-unwind\" ");
-                    } else {
-                        output.push_str("extern \"cdecl\" ");
-                    }
-                }
-                Abi::Stdcall { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"stdcall-unwind\" ");
-                    } else {
-                        output.push_str("extern \"stdcall\" ");
-                    }
-                }
-                Abi::Fastcall { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"fastcall-unwind\" ");
-                    } else {
-                        output.push_str("extern \"fastcall\" ");
-                    }
-                }
-                Abi::Aapcs { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"aapcs-unwind\" ");
-                    } else {
-                        output.push_str("extern \"aapcs\" ");
-                    }
-                }
-                Abi::Win64 { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"win64-unwind\" ");
-                    } else {
-                        output.push_str("extern \"win64\" ");
-                    }
-                }
-                Abi::SysV64 { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"sysv64-unwind\" ");
-                    } else {
-                        output.push_str("extern \"sysv64\" ");
-                    }
-                }
-                Abi::System { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"system-unwind\" ");
-                    } else {
-                        output.push_str("extern \"system\" ");
-                    }
-                }
-                Abi::Other(abi) => {
-                    output.push_str(&format!("extern \"{}\" ", abi));
-                }
-            }
-
-            // Function name
-            if let Some(name) = &item.name {
-                output.push_str(&format!("fn {}", name));
-
-                // Generic parameters
-                format_generics(output, &function.generics, data);
-
-                // Parameters
-                output.push('(');
-                for (i, (param_name, param_type)) in function.sig.inputs.iter().enumerate() {
-                    output.push_str(&format!(
-                        "{}: {}",
-                        param_name,
-                        format_type(param_type, data)
-                    ));
-                    if i < function.sig.inputs.len() - 1 || function.sig.is_c_variadic {
-                        output.push_str(", ");
-                    }
-                }
-
-                // Variadic
-                if function.sig.is_c_variadic {
-                    output.push_str("...");
-                }
-
-                output.push(')');
-
-                // Return type
-                if let Some(return_type) = &function.sig.output {
-                    output.push_str(&format!(" -> {}", format_type(return_type, data)));
-                }
-
-                // Where clause
-                format_where_clause(output, &function.generics.where_predicates, data);
-
-                // Function body indication
-                if function.has_body {
-                    output.push_str(" { /* ... */ }");
-                } else {
-                    output.push(';');
-                }
-            }
-        }
-        ItemEnum::Trait(trait_) => {
-            // Trait modifiers
-            if trait_.is_auto {
-                output.push_str("auto ");
-            }
-            if trait_.is_unsafe {
-                output.push_str("unsafe ");
-            }
-
-            // Trait definition
-            if let Some(name) = &item.name {
-                output.push_str(&format!("trait {}", name));
-                format_generics(output, &trait_.generics, data);
-
-                // Trait bounds
-                if !trait_.bounds.is_empty() {
-                    output.push_str(": ");
-                    format_bounds(output, &trait_.bounds, data);
-                }
-
-                // Where clause
-                format_where_clause(output, &trait_.generics.where_predicates, data);
-
-                output.push_str(" {\n    /* Associated items */\n}");
-            }
-        }
-        ItemEnum::TraitAlias(trait_alias) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("trait {}", name));
-                format_generics(output, &trait_alias.generics, data);
-                output.push_str(" = ");
-                format_bounds(output, &trait_alias.params, data);
-                format_where_clause(output, &trait_alias.generics.where_predicates, data);
-                output.push(';');
-            }
-        }
-        ItemEnum::Impl(impl_) => {
-            // Impl modifiers
-            if impl_.is_unsafe {
-                output.push_str("unsafe ");
-            }
-
-            output.push_str("impl");
-
-            // Generics
-            format_generics(output, &impl_.generics, data);
-
-            // Trait reference if this is a trait impl
-            if let Some(trait_) = &impl_.trait_ {
-                if impl_.is_negative {
-                    output.push_str(" !");
-                } else {
-                    output.push(' ');
-                }
-
-                output.push_str(&trait_.path);
-                if let Some(args) = &trait_.args {
-                    let mut args_str = String::new();
-                    format_generic_args(&mut args_str, args, data);
-                    output.push_str(&args_str);
-                }
-
-                output.push_str(" for ");
-            }
-
-            // For type
-            output.push_str(&format_type(&impl_.for_, data));
-
-            // Where clause
-            format_where_clause(output, &impl_.generics.where_predicates, data);
-
-            output.push_str(" {\n    /* Associated items */\n}");
-
-            // Add note if this is a compiler-generated impl
-            if impl_.is_synthetic {
-                output.push_str("\n// Note: This impl is compiler-generated");
-            }
-        }
-        ItemEnum::TypeAlias(type_alias) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("type {}", name));
-                format_generics(output, &type_alias.generics, data);
-                format_where_clause(output, &type_alias.generics.where_predicates, data);
-                output.push_str(&format!(" = {};", format_type(&type_alias.type_, data)));
-            }
-        }
-        ItemEnum::Constant { type_, const_ } => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!(
-                    "const {}: {} = {};",
-                    name,
-                    format_type(type_, data),
-                    const_.expr
-                ));
-            }
-        }
-        ItemEnum::Static(static_) => {
-            if let Some(name) = &item.name {
-                output.push_str("static ");
-                if static_.is_mutable {
-                    output.push_str("mut ");
-                }
-                if static_.is_unsafe {
-                    output.push_str("/* unsafe */ ");
-                }
-                output.push_str(&format!(
-                    "{}: {} = {};",
-                    name,
-                    format_type(&static_.type_, data),
-                    static_.expr
-                ));
-            }
-        }
-        ItemEnum::Macro(macro_body) => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!(
-                    "macro_rules! {} {{\n    /* {} */\n}}",
-                    name, macro_body
-                ));
-            }
-        }
-        ItemEnum::ProcMacro(proc_macro) => {
-            if let Some(name) = &item.name {
-                output.push_str("#[proc_macro");
-                match proc_macro.kind {
-                    MacroKind::Bang => output.push(']'),
-
-                    MacroKind::Attr => output.push_str("_attribute]"),
-                    MacroKind::Derive => {
-                        output.push_str("_derive]");
-                        if !proc_macro.helpers.is_empty() {
-                            output.push_str("\n// Helpers: ");
-                            for (i, helper) in proc_macro.helpers.iter().enumerate() {
-                                output.push_str(&format!("#[{}]", helper));
-                                if i < proc_macro.helpers.len() - 1 {
-                                    output.push_str(", ");
-                                }
-                            }
-                        }
-                    }
-                }
-                output.push_str(&format!(
-                    "\npub fn {}(/* ... */) -> /* ... */ {{\n    /* ... */\n}}",
-                    name
-                ));
-            }
-        }
-        ItemEnum::ExternCrate { name, rename } => {
-            output.push_str(&format!("extern crate {}", name));
-            if let Some(rename_val) = rename {
-                output.push_str(&format!(" as {}", rename_val));
-            }
-            output.push(';');
-        }
-        ItemEnum::Use(use_item) => {
-            output.push_str(&format!("use {}", use_item.source));
-            if use_item.is_glob {
-                output.push_str("::*");
-            } else if use_item.name
-                != use_item
-                    .source
-                    .split("::")
-                    .last()
-                    .unwrap_or(&use_item.source)
-            {
-                output.push_str(&format!(" as {}", use_item.name));
-            }
-            output.push(';');
-        }
-        ItemEnum::StructField(field_type) => {
-            // For struct fields, just output the type
-            if let Some(name) = &item.name {
-                match &item.visibility {
-                    Visibility::Public => output.push_str("pub "),
-                    Visibility::Crate => output.push_str("pub(crate) "),
-                    Visibility::Restricted { path, .. } => {
-                        output.push_str(&format!("pub(in {}) ", path))
-                    }
-                    Visibility::Default => {}
-                }
-                output.push_str(&format!("{}: {}", name, format_type(field_type, data)));
-            } else {
-                output.push_str(&format_type(field_type, data));
-            }
-        }
-        ItemEnum::Variant(variant) => {
-            // For enum variants
-            if let Some(name) = &item.name {
-                output.push_str(name);
-
-                match &variant.kind {
-                    VariantKind::Plain => {}
-                    VariantKind::Tuple(fields) => {
-                        output.push('(');
-                        for (i, field_opt) in fields.iter().enumerate() {
-                            if let Some(field_id) = field_opt {
-                                if let Some(field_item) = data.index.get(field_id) {
-                                    if let ItemEnum::StructField(field_type) = &field_item.inner {
-                                        output.push_str(&format_type(field_type, data));
-                                    }
-                                }
-                                if i < fields.len() - 1 {
-                                    output.push_str(", ");
-                                }
-                            } else {
-                                // For stripped fields
-                                output.push_str("/* private field */");
-                                if i < fields.len() - 1 {
-                                    output.push_str(", ");
-                                }
-                            }
-                        }
-                        output.push(')');
-                    }
-                    VariantKind::Struct {
-                        fields,
-                        has_stripped_fields,
-                    } => {
-                        output.push_str(" {\n");
-                        for &field_id in fields {
-                            if let Some(field_item) = data.index.get(&field_id) {
-                                if let Some(field_name) = &field_item.name {
-                                    if let ItemEnum::StructField(field_type) = &field_item.inner {
-                                        output.push_str(&format!(
-                                            "    {}: {},\n",
-                                            field_name,
-                                            format_type(field_type, data)
-                                        ));
-                                    }
-                                }
-                            }
-                        }
-                        if *has_stripped_fields {
-                            output.push_str("    // Some fields omitted\n");
-                        }
-                        output.push('}');
-                    }
-                }
-
-                if let Some(discriminant) = &variant.discriminant {
-                    output.push_str(&format!(" = {}", discriminant.expr));
-                }
-            }
-        }
-        ItemEnum::Primitive(primitive) => {
-            output.push_str(&format!("// Primitive type: {}", primitive.name));
-        }
-        ItemEnum::ExternType => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("extern {{ type {}; }}", name));
-            }
-        }
-        ItemEnum::AssocConst { type_, value } => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("const {}: {}", name, format_type(type_, data)));
-                if let Some(val) = value {
-                    output.push_str(&format!(" = {}", val));
-                }
-                output.push(';');
-            }
-        }
-        ItemEnum::AssocType {
-            generics,
-            bounds,
-            type_,
-        } => {
-            if let Some(name) = &item.name {
-                output.push_str(&format!("type {}", name));
-                format_generics(output, generics, data);
-
-                if !bounds.is_empty() {
-                    output.push_str(": ");
-                    format_bounds(output, bounds, data);
-                }
-
-                if let Some(ty) = type_ {
-                    output.push_str(&format!(" = {}", format_type(ty, data)));
-                }
-
-                format_where_clause(output, &generics.where_predicates, data);
-                output.push(';');
-            }
-        }
-    }
-}
-
-fn format_generics(output: &mut String, generics: &Generics, data: &Crate) {
-    if generics.params.is_empty() {
-        return;
-    }
-
-    output.push('<');
-    for (i, param) in generics.params.iter().enumerate() {
-        match &param.kind {
-            GenericParamDefKind::Lifetime { outlives } => {
-                output.push_str(&format!("'{}", param.name));
-                if !outlives.is_empty() {
-                    output.push_str(": ");
-                    for (j, lifetime) in outlives.iter().enumerate() {
-                        output.push_str(&format!("'{}", lifetime));
-                        if j < outlives.len() - 1 {
-                            output.push_str(" + ");
-                        }
-                    }
-                }
-            }
-            GenericParamDefKind::Type {
-                bounds,
-                default,
-                is_synthetic,
-            } => {
-                // If synthetic, add a note
-                if *is_synthetic {
-                    output.push_str("/* synthetic */ ");
-                }
-
-                output.push_str(&param.name);
-                if !bounds.is_empty() {
-                    output.push_str(": ");
-                    format_bounds(output, bounds, data);
-                }
-                if let Some(default_type) = default {
-                    output.push_str(&format!(" = {}", format_type(default_type, data)));
-                }
-            }
-            GenericParamDefKind::Const { type_, default } => {
-                output.push_str(&format!(
-                    "const {}: {}",
-                    param.name,
-                    format_type(type_, data)
-                ));
-                if let Some(default_value) = default {
-                    output.push_str(&format!(" = {}", default_value));
-                }
-            }
-        }
-
-        if i < generics.params.len() - 1 {
-            output.push_str(", ");
-        }
-    }
-    output.push('>');
-}
-
-fn format_where_clause(output: &mut String, predicates: &[WherePredicate], data: &Crate) {
-    if predicates.is_empty() {
-        return;
-    }
-
-    output.push_str("\nwhere\n    ");
-    for (i, predicate) in predicates.iter().enumerate() {
-        match predicate {
-            WherePredicate::BoundPredicate {
-                type_,
-                bounds,
-                generic_params,
-            } => {
-                if !generic_params.is_empty() {
-                    output.push_str("for<");
-                    for (j, param) in generic_params.iter().enumerate() {
-                        match &param.kind {
-                            GenericParamDefKind::Lifetime { .. } => {
-                                output.push_str(&format!("'{}", param.name));
-                            }
-                            _ => output.push_str(&param.name),
-                        }
-
-                        if j < generic_params.len() - 1 {
-                            output.push_str(", ");
-                        }
-                    }
-                    output.push_str("> ");
-                }
-
-                output.push_str(&format_type(type_, data));
-
-                if !bounds.is_empty() {
-                    output.push_str(": ");
-                    format_bounds(output, bounds, data);
-                }
-            }
-            WherePredicate::LifetimePredicate { lifetime, outlives } => {
-                output.push_str(&format!("'{}", lifetime));
-                if !outlives.is_empty() {
-                    output.push_str(": ");
-                    for (j, outlive) in outlives.iter().enumerate() {
-                        output.push_str(&format!("'{}", outlive));
-                        if j < outlives.len() - 1 {
-                            output.push_str(" + ");
-                        }
-                    }
-                }
-            }
-            WherePredicate::EqPredicate { lhs, rhs } => {
-                output.push_str(&format_type(lhs, data));
-                output.push_str(" = ");
-                match rhs {
-                    Term::Type(type_) => output.push_str(&format_type(type_, data)),
-                    Term::Constant(constant) => output.push_str(&constant.expr),
-                }
-            }
-        }
-
-        if i < predicates.len() - 1 {
-            output.push_str(",\n    ");
-        }
-    }
-}
-
-fn format_bounds(output: &mut String, bounds: &[GenericBound], data: &Crate) {
-    for (i, bound) in bounds.iter().enumerate() {
-        match bound {
-            GenericBound::TraitBound {
-                trait_,
-                generic_params,
-                modifier,
-            } => {
-                match modifier {
-                    TraitBoundModifier::None => {}
-                    TraitBoundModifier::Maybe => output.push('?'),
-                    TraitBoundModifier::MaybeConst => output.push_str("~const "),
-                }
-
-                if !generic_params.is_empty() {
-                    output.push_str("for<");
-                    for (j, param) in generic_params.iter().enumerate() {
-                        match &param.kind {
-                            GenericParamDefKind::Lifetime { .. } => {
-                                output.push_str(&format!("'{}", param.name));
-                            }
-                            _ => output.push_str(&param.name),
-                        }
-
-                        if j < generic_params.len() - 1 {
-                            output.push_str(", ");
-                        }
-                    }
-                    output.push_str("> ");
-                }
-
-                output.push_str(&trait_.path);
-                if let Some(args) = &trait_.args {
-                    let mut args_str = String::new();
-                    format_generic_args(&mut args_str, args, data);
-                    output.push_str(&args_str);
-                }
-            }
-            GenericBound::Outlives(lifetime) => {
-                output.push_str(&format!("'{}", lifetime));
-            }
-            GenericBound::Use(args) => {
-                output.push_str("use<");
-                for (i, arg) in args.iter().enumerate() {
-                    match arg {
-                        PreciseCapturingArg::Lifetime(lifetime) => {
-                            output.push_str(&format!("'{}", lifetime))
-                        }
-                        PreciseCapturingArg::Param(param) => output.push_str(param),
-                    }
-
-                    if i < args.len() - 1 {
-                        output.push_str(", ");
-                    }
-                }
-                output.push('>');
-            }
-        }
-
-        if i < bounds.len() - 1 {
-            output.push_str(" + ");
-        }
-    }
-}
-
-fn format_generic_args(output: &mut String, args: &GenericArgs, data: &Crate) {
-    match args {
-        GenericArgs::AngleBracketed { args, constraints } => {
-            if args.is_empty() && constraints.is_empty() {
-                return;
-            }
-
-            output.push('<');
-
-            // Format args
-            for (i, arg) in args.iter().enumerate() {
-                match arg {
-                    GenericArg::Lifetime(lifetime) => output.push_str(&format!("'{}", lifetime)),
-                    GenericArg::Type(type_) => output.push_str(&format_type(type_, data)),
-                    GenericArg::Const(constant) => output.push_str(&constant.expr),
-                    GenericArg::Infer => output.push('_'),
-                }
-
-                if i < args.len() - 1 || !constraints.is_empty() {
-                    output.push_str(", ");
-                }
-            }
-
-            // Format constraints
-            for (i, constraint) in constraints.iter().enumerate() {
-                output.push_str(&constraint.name.to_string());
-
-                // Format constraint args if present
-                let mut args_str = String::new();
-                format_generic_args(&mut args_str, &constraint.args, data);
-                if !args_str.is_empty() && args_str != "<>" {
-                    output.push_str(&args_str);
-                }
-
-                match &constraint.binding {
-                    AssocItemConstraintKind::Equality(term) => {
-                        output.push_str(" = ");
-                        match term {
-                            Term::Type(type_) => output.push_str(&format_type(type_, data)),
-                            Term::Constant(constant) => output.push_str(&constant.expr),
-                        }
-                    }
-                    AssocItemConstraintKind::Constraint(bounds) => {
-                        output.push_str(": ");
-                        format_bounds(output, bounds, data);
-                    }
-                }
-
-                if i < constraints.len() - 1 {
-                    output.push_str(", ");
-                }
-            }
-
-            output.push('>');
-        }
-        GenericArgs::Parenthesized {
-            inputs,
-            output: output_type,
-        } => {
-            output.push('(');
-
-            for (i, input) in inputs.iter().enumerate() {
-                output.push_str(&format_type(input, data));
-                if i < inputs.len() - 1 {
-                    output.push_str(", ");
-                }
-            }
-
-            output.push(')');
-
-            if let Some(output_ty) = output_type {
-                output.push_str(&format!(" -> {}", format_type(output_ty, data)));
-            }
-        }
-        GenericArgs::ReturnTypeNotation => {
-            output.push_str("::method(..)");
-        }
-    }
-}
-
-fn format_type(ty: &Type, data: &Crate) -> String {
-    let mut output = String::new();
-
-    match ty {
-        Type::ResolvedPath(path) => {
-            output.push_str(&path.path);
-            if let Some(args) = &path.args {
-                let mut args_str = String::new();
-                format_generic_args(&mut args_str, args, data);
-                output.push_str(&args_str);
-            }
-        }
-        Type::DynTrait(dyn_trait) => {
-            output.push_str("dyn ");
-
-            for (i, trait_) in dyn_trait.traits.iter().enumerate() {
-                // Higher-rank bounds if necessary
-                if !trait_.generic_params.is_empty() {
-                    output.push_str("for<");
-                    for (j, param) in trait_.generic_params.iter().enumerate() {
-                        match &param.kind {
-                            GenericParamDefKind::Lifetime { .. } => {
-                                output.push_str(&format!("'{}", param.name));
-                            }
-                            _ => output.push_str(&param.name),
-                        }
-
-                        if j < trait_.generic_params.len() - 1 {
-                            output.push_str(", ");
-                        }
-                    }
-                    output.push_str("> ");
-                }
-
-                output.push_str(&trait_.trait_.path);
-                if let Some(args) = &trait_.trait_.args {
-                    let mut args_str = String::new();
-                    format_generic_args(&mut args_str, args, data);
-                    output.push_str(&args_str);
-                }
-
-                if i < dyn_trait.traits.len() - 1 {
-                    output.push_str(" + ");
-                }
-            }
-
-            // Lifetime bound if present
-            if let Some(lifetime) = &dyn_trait.lifetime {
-                output.push_str(&format!(" + '{}", lifetime));
-            }
-        }
-        Type::Generic(name) => {
-            output.push_str(name);
-        }
-        Type::Primitive(name) => {
-            output.push_str(name);
-        }
-        Type::FunctionPointer(fn_ptr) => {
-            // For clarity about the parameters
-            if !fn_ptr.generic_params.is_empty() {
-                output.push_str("for<");
-                for (j, param) in fn_ptr.generic_params.iter().enumerate() {
-                    match &param.kind {
-                        GenericParamDefKind::Lifetime { .. } => {
-                            output.push_str(&format!("'{}", param.name));
-                        }
-                        _ => output.push_str(&param.name),
-                    }
-
-                    if j < fn_ptr.generic_params.len() - 1 {
-                        output.push_str(", ");
-                    }
-                }
-                output.push_str("> ");
-            }
-
-            // Function header (const, unsafe, extern, etc.)
-            if fn_ptr.header.is_const {
-                output.push_str("const ");
-            }
-            if fn_ptr.header.is_unsafe {
-                output.push_str("unsafe ");
-            }
-
-            // ABI
-            match &fn_ptr.header.abi {
-                Abi::Rust => {}
-                Abi::C { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"C-unwind\" ");
-                    } else {
-                        output.push_str("extern \"C\" ");
-                    }
-                }
-                Abi::Cdecl { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"cdecl-unwind\" ");
-                    } else {
-                        output.push_str("extern \"cdecl\" ");
-                    }
-                }
-                Abi::Stdcall { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"stdcall-unwind\" ");
-                    } else {
-                        output.push_str("extern \"stdcall\" ");
-                    }
-                }
-                Abi::Fastcall { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"fastcall-unwind\" ");
-                    } else {
-                        output.push_str("extern \"fastcall\" ");
-                    }
-                }
-                Abi::Aapcs { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"aapcs-unwind\" ");
-                    } else {
-                        output.push_str("extern \"aapcs\" ");
-                    }
-                }
-                Abi::Win64 { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"win64-unwind\" ");
-                    } else {
-                        output.push_str("extern \"win64\" ");
-                    }
-                }
-                Abi::SysV64 { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"sysv64-unwind\" ");
-                    } else {
-                        output.push_str("extern \"sysv64\" ");
-                    }
-                }
-                Abi::System { unwind } => {
-                    if *unwind {
-                        output.push_str("extern \"system-unwind\" ");
-                    } else {
-                        output.push_str("extern \"system\" ");
-                    }
-                }
-                Abi::Other(abi) => {
-                    output.push_str(&format!("extern \"{}\" ", abi));
-                }
-            }
-
-            output.push_str("fn(");
-
-            // Parameters
-            for (i, (_, param_type)) in fn_ptr.sig.inputs.iter().enumerate() {
-                output.push_str(&format_type(param_type, data));
-                if i < fn_ptr.sig.inputs.len() - 1 || fn_ptr.sig.is_c_variadic {
-                    output.push_str(", ");
-                }
-            }
-
-            // Variadic
-            if fn_ptr.sig.is_c_variadic {
-                output.push_str("...");
-            }
-
-            output.push(')');
-
-            // Return type
-            if let Some(return_type) = &fn_ptr.sig.output {
-                output.push_str(&format!(" -> {}", format_type(return_type, data)));
-            }
-        }
-        Type::Tuple(types) => {
-            if types.is_empty() {
-                output.push_str("()");
-            } else {
-                output.push('(');
-                for (i, ty) in types.iter().enumerate() {
-                    output.push_str(&format_type(ty, data));
-                    if i < types.len() - 1 {
-                        output.push_str(", ");
-                    }
-                }
-                output.push(')');
-            }
-        }
-        Type::Slice(ty) => {
-            output.push_str(&format!("[{}]", format_type(ty, data)));
-        }
-        Type::Array { type_, len } => {
-            output.push_str(&format!("[{}; {}]", format_type(type_, data), len));
-        }
-        Type::Pat {
-            type_,
-            __pat_unstable_do_not_use,
-        } => {
-            output.push_str(&format!(
-                "{} is {}",
-                format_type(type_, data),
-                __pat_unstable_do_not_use
-            ));
-        }
-        Type::ImplTrait(bounds) => {
-            output.push_str("impl ");
-
-            let mut bounds_str = String::new();
-            format_bounds(&mut bounds_str, bounds, data);
-            output.push_str(&bounds_str);
-        }
-        Type::Infer => {
-            output.push('_');
-        }
-        Type::RawPointer { is_mutable, type_ } => {
-            if *is_mutable {
-                output.push_str("*mut ");
-            } else {
-                output.push_str("*const ");
-            }
-            output.push_str(&format_type(type_, data));
-        }
-        Type::BorrowedRef {
-            lifetime,
-            is_mutable,
-            type_,
-        } => {
-            output.push('&');
-            if let Some(lt) = lifetime {
-                output.push_str(&format!("'{} ", lt));
-            }
-            if *is_mutable {
-                output.push_str("mut ");
-            }
-            output.push_str(&format_type(type_, data));
-        }
-        Type::QualifiedPath {
-            name,
-            args,
-            self_type,
-            trait_,
-        } => {
-            output.push('<');
-            output.push_str(&format_type(self_type, data));
-
-            if let Some(trait_path) = trait_ {
-                output.push_str(&format!(" as {}", trait_path.path));
-                if let Some(trait_args) = &trait_path.args {
-                    let mut args_str = String::new();
-                    format_generic_args(&mut args_str, trait_args, data);
-                    output.push_str(&args_str);
-                }
-            }
-
-            output.push_str(&format!(">::{}", name));
-
-            let mut args_str = String::new();
-            format_generic_args(&mut args_str, args, data);
-            if args_str != "<>" && !args_str.is_empty() {
-                output.push_str(&args_str);
-            }
-        }
-    }
-
-    output
-}
-
-fn process_module_details(output: &mut String, module: &Module, data: &Crate, _level: usize) {
-    if module.is_stripped {
-        output.push_str(
-            "> **Note:** This module is marked as stripped. Some items may be omitted.\n\n",
-        );
-    }
-
-    // Reset level when entering a module to avoid excessive nesting
-    // This ensures that module contents are always at a reasonable heading level
-    process_items(output, &module.items, data, 3);
-}
-
-fn process_struct_details(output: &mut String, struct_: &Struct, data: &Crate, level: usize) {
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-    // Detail fields based on struct kind
-    match &struct_.kind {
-        StructKind::Unit => {
-            // Nothing to detail for unit structs
-        }
-        StructKind::Tuple(fields) => {
-            // Use heading_level for Fields section (since level is already incremented in process_item)
-            output.push_str(&format!("{} Fields\n\n", "#".repeat(heading_level)));
-            output.push_str("| Index | Type | Documentation |\n");
-            output.push_str("|-------|------|---------------|\n");
-
-            for (i, field_opt) in fields.iter().enumerate() {
-                if let Some(field_id) = field_opt {
-                    if let Some(field_item) = data.index.get(field_id) {
-                        if let ItemEnum::StructField(field_type) = &field_item.inner {
-                            let docs = field_item
-                                .docs
-                                .as_deref()
-                                .unwrap_or("")
-                                .replace("\n", "<br>");
-                            output.push_str(&format!(
-                                "| {} | `{}` | {} |\n",
-                                i,
-                                format_type(field_type, data),
-                                docs
-                            ));
-                        }
-                    }
-                } else {
-                    output.push_str(&format!("| {} | `private` | *Private field* |\n", i));
-                }
-            }
-            output.push('\n');
-        }
-        StructKind::Plain {
-            fields,
-            has_stripped_fields,
-        } => {
-            // Use heading_level for Fields section
-            output.push_str(&format!("{} Fields\n\n", "#".repeat(heading_level)));
-            output.push_str("| Name | Type | Documentation |\n");
-            output.push_str("|------|------|---------------|\n");
-
-            for &field_id in fields {
-                if let Some(field_item) = data.index.get(&field_id) {
-                    if let Some(field_name) = &field_item.name {
-                        if let ItemEnum::StructField(field_type) = &field_item.inner {
-                            let docs = field_item
-                                .docs
-                                .as_deref()
-                                .unwrap_or("")
-                                .replace("\n", "<br>");
-                            output.push_str(&format!(
-                                "| `{}` | `{}` | {} |\n",
-                                field_name,
-                                format_type(field_type, data),
-                                docs
-                            ));
-                        }
-                    }
-                }
-            }
-
-            if *has_stripped_fields {
-                output.push_str("| *private fields* | ... | *Some fields have been omitted* |\n");
-            }
-
-            output.push('\n');
-        }
-    }
-
-    // Process impls
-    if !struct_.impls.is_empty() {
-        // Use heading_level for Implementations section
-        output.push_str(&format!(
-            "{} Implementations\n\n",
-            "#".repeat(heading_level)
-        ));
-
-        // Group impls by trait
-        let mut trait_impls: std::collections::HashMap<String, Vec<Id>> =
-            std::collections::HashMap::new();
-        let mut inherent_impls: Vec<Id> = Vec::new();
-
-        for &impl_id in &struct_.impls {
-            if let Some(impl_item) = data.index.get(&impl_id) {
-                if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                    if let Some(trait_) = &impl_.trait_ {
-                        let trait_name = trait_.path.clone();
-                        trait_impls.entry(trait_name).or_default().push(impl_id);
-                    } else {
-                        // Inherent impl
-                        inherent_impls.push(impl_id);
-                    }
-                }
-            }
-        }
-
-        // First list inherent impls
-        if !inherent_impls.is_empty() {
-            // Use level+1 for Methods (one level deeper than Implementations)
-            output.push_str(&format!(
-                "{} Methods\n\n",
-                "#".repeat(std::cmp::min(heading_level + 1, 6))
-            ));
-            for &impl_id in &inherent_impls {
-                if let Some(impl_item) = data.index.get(&impl_id) {
-                    if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                        for &item_id in &impl_.items {
-                            if let Some(method_item) = data.index.get(&item_id) {
-                                if let ItemEnum::Function(_) = &method_item.inner {
-                                    // Format method signature
-                                    let mut method_signature = String::new();
-                                    format_item_signature(&mut method_signature, method_item, data);
-
-                                    // Output with proper code block formatting
-                                    output.push_str("- ```rust\n  ");
-                                    output.push_str(&method_signature.trim());
-                                    output.push_str("\n  ```");
-
-                                    // Add documentation if available
-                                    if let Some(docs) = &method_item.docs {
-                                        if let Some(first_line) = docs.lines().next() {
-                                            if !first_line.trim().is_empty() {
-                                                output.push_str(&format!("\n  {}", first_line));
-                                            }
-                                        }
-                                    }
-                                    output.push_str("\n\n");
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Then list trait impls
-        if !trait_impls.is_empty() {
-            // Use level+1 for Trait Implementations (one level deeper than Implementations)
-            output.push_str(&format!(
-                "{} Trait Implementations\n\n",
-                "#".repeat(std::cmp::min(heading_level + 1, 6))
-            ));
-            for (trait_name, impls) in trait_impls {
-                output.push_str(&format!("- **{}**\n", trait_name));
-                for &impl_id in &impls {
-                    if let Some(impl_item) = data.index.get(&impl_id) {
-                        if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                            for &item_id in &impl_.items {
-                                if let Some(method_item) = data.index.get(&item_id) {
-                                    if let ItemEnum::Function(_) = &method_item.inner {
-                                        // Format method signature
-                                        let mut method_signature = String::new();
-                                        format_item_signature(
-                                            &mut method_signature,
-                                            method_item,
-                                            data,
-                                        );
-
-                                        // Output with proper code block formatting
-                                        output.push_str("  - ```rust\n    ");
-                                        output.push_str(&method_signature.trim());
-                                        output.push_str("\n    ```");
-
-                                        // Add documentation if available
-                                        if let Some(docs) = &method_item.docs {
-                                            if let Some(first_line) = docs.lines().next() {
-                                                if !first_line.trim().is_empty() {
-                                                    output
-                                                        .push_str(&format!("\n    {}", first_line));
-                                                }
-                                            }
-                                        }
-                                        output.push_str("\n\n");
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn process_enum_details(output: &mut String, enum_: &Enum, data: &Crate, level: usize) {
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-    // Detail variants with proper nesting
-    output.push_str(&format!("{} Variants\n\n", "#".repeat(heading_level)));
-
-    for &variant_id in &enum_.variants {
-        if let Some(variant_item) = data.index.get(&variant_id) {
-            if let Some(variant_name) = &variant_item.name {
-                // Use heading_level + 1 for individual variants (capped at 6)
-                let variant_heading_level = std::cmp::min(heading_level + 1, 6);
-                output.push_str(&format!(
-                    "{} `{}`\n\n",
-                    "#".repeat(variant_heading_level),
-                    variant_name
-                ));
-
-                // Add variant docs if available
-                if let Some(docs) = &variant_item.docs {
-                    output.push_str(&format!("{}\n\n", docs));
-                }
-
-                if let ItemEnum::Variant(variant) = &variant_item.inner {
-                    match &variant.kind {
-                        VariantKind::Plain => {
-                            // Nothing additional to display for plain variants
-                            if let Some(discriminant) = &variant.discriminant {
-                                output.push_str(&format!(
-                                    "Discriminant: `{}`\n\n",
-                                    discriminant.expr
-                                ));
-                            }
-                        }
-                        VariantKind::Tuple(fields) => {
-                            output.push_str("Fields:\n\n");
-                            output.push_str("| Index | Type | Documentation |\n");
-                            output.push_str("|-------|------|---------------|\n");
-
-                            for (i, field_opt) in fields.iter().enumerate() {
-                                if let Some(field_id) = field_opt {
-                                    if let Some(field_item) = data.index.get(field_id) {
-                                        if let ItemEnum::StructField(field_type) = &field_item.inner
-                                        {
-                                            let docs = field_item
-                                                .docs
-                                                .as_deref()
-                                                .unwrap_or("")
-                                                .replace("\n", "<br>");
-                                            output.push_str(&format!(
-                                                "| {} | `{}` | {} |\n",
-                                                i,
-                                                format_type(field_type, data),
-                                                docs
-                                            ));
-                                        }
-                                    }
-                                } else {
-                                    output.push_str(&format!(
-                                        "| {} | `private` | *Private field* |\n",
-                                        i
-                                    ));
-                                }
-                            }
-                            output.push('\n');
-                        }
-                        VariantKind::Struct {
-                            fields,
-                            has_stripped_fields,
-                        } => {
-                            output.push_str("Fields:\n\n");
-                            output.push_str("| Name | Type | Documentation |\n");
-                            output.push_str("|------|------|---------------|\n");
-
-                            for &field_id in fields {
-                                if let Some(field_item) = data.index.get(&field_id) {
-                                    if let Some(field_name) = &field_item.name {
-                                        if let ItemEnum::StructField(field_type) = &field_item.inner
-                                        {
-                                            let docs = field_item
-                                                .docs
-                                                .as_deref()
-                                                .unwrap_or("")
-                                                .replace("\n", "<br>");
-                                            output.push_str(&format!(
-                                                "| `{}` | `{}` | {} |\n",
-                                                field_name,
-                                                format_type(field_type, data),
-                                                docs
-                                            ));
-                                        }
-                                    }
-                                }
-                            }
-
-                            if *has_stripped_fields {
-                                output.push_str("| *private fields* | ... | *Some fields have been omitted* |\n");
-                            }
-
-                            output.push('\n');
-                        }
-                    }
-
-                    if let Some(discriminant) = &variant.discriminant {
-                        output
-                            .push_str(&format!("Discriminant value: `{}`\n\n", discriminant.value));
-                    }
-                }
-            }
-        }
-    }
-
-    if enum_.has_stripped_variants {
-        output.push_str(
-            "*Note: Some variants have been omitted because they are private or hidden.*\n\n",
-        );
-    }
-
-    // Process impls (same as for struct)
-    if !enum_.impls.is_empty() {
-        output.push_str(&format!(
-            "{} Implementations\n\n",
-            "#".repeat(heading_level)
-        ));
-
-        // Group impls by trait
-        let mut trait_impls: std::collections::HashMap<String, Vec<Id>> =
-            std::collections::HashMap::new();
-        let mut inherent_impls: Vec<Id> = Vec::new();
-
-        for &impl_id in &enum_.impls {
-            if let Some(impl_item) = data.index.get(&impl_id) {
-                if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                    if let Some(trait_) = &impl_.trait_ {
-                        let trait_name = trait_.path.clone();
-                        trait_impls.entry(trait_name).or_default().push(impl_id);
-                    } else {
-                        // Inherent impl
-                        inherent_impls.push(impl_id);
-                    }
-                }
-            }
-        }
-
-        // First list inherent impls
-        if !inherent_impls.is_empty() {
-            let methods_level = std::cmp::min(heading_level + 1, 6);
-            output.push_str(&format!("{} Methods\n\n", "#".repeat(methods_level)));
-            for &impl_id in &inherent_impls {
-                if let Some(impl_item) = data.index.get(&impl_id) {
-                    if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                        for &item_id in &impl_.items {
-                            if let Some(method_item) = data.index.get(&item_id) {
-                                if let ItemEnum::Function(_) = &method_item.inner {
-                                    // Format method signature
-                                    let mut method_signature = String::new();
-                                    format_item_signature(&mut method_signature, method_item, data);
-
-                                    // Output with proper code block formatting
-                                    output.push_str("- ```rust\n  ");
-                                    output.push_str(&method_signature.trim());
-                                    output.push_str("\n  ```");
-
-                                    // Add documentation if available
-                                    if let Some(docs) = &method_item.docs {
-                                        if let Some(first_line) = docs.lines().next() {
-                                            if !first_line.trim().is_empty() {
-                                                output.push_str(&format!("\n  {}", first_line));
-                                            }
-                                        }
-                                    }
-                                    output.push_str("\n\n");
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Then list trait impls
-        if !trait_impls.is_empty() {
-            let trait_impl_level = std::cmp::min(heading_level + 1, 6);
-            output.push_str(&format!(
-                "{} Trait Implementations\n\n",
-                "#".repeat(trait_impl_level)
-            ));
-            for (trait_name, impls) in trait_impls {
-                output.push_str(&format!("- **{}**\n", trait_name));
-                for &impl_id in &impls {
-                    if let Some(impl_item) = data.index.get(&impl_id) {
-                        if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                            for &item_id in &impl_.items {
-                                if let Some(method_item) = data.index.get(&item_id) {
-                                    if let ItemEnum::Function(_) = &method_item.inner {
-                                        // Format method signature
-                                        let mut method_signature = String::new();
-                                        format_item_signature(
-                                            &mut method_signature,
-                                            method_item,
-                                            data,
-                                        );
-
-                                        // Output with proper code block formatting
-                                        output.push_str("  - ```rust\n    ");
-                                        output.push_str(&method_signature.trim());
-                                        output.push_str("\n    ```");
-
-                                        // Add documentation if available
-                                        if let Some(docs) = &method_item.docs {
-                                            if let Some(first_line) = docs.lines().next() {
-                                                if !first_line.trim().is_empty() {
-                                                    output
-                                                        .push_str(&format!("\n    {}", first_line));
-                                                }
-                                            }
-                                        }
-                                        output.push_str("\n\n");
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn process_union_details(output: &mut String, union_: &Union, data: &Crate, level: usize) {
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-    // Detail fields
-    output.push_str(&format!("{} Fields\n\n", "#".repeat(heading_level)));
-    output.push_str("| Name | Type | Documentation |\n");
-    output.push_str("|------|------|---------------|\n");
-
-    for &field_id in &union_.fields {
-        if let Some(field_item) = data.index.get(&field_id) {
-            if let Some(field_name) = &field_item.name {
-                if let ItemEnum::StructField(field_type) = &field_item.inner {
-                    let docs = field_item
-                        .docs
-                        .as_deref()
-                        .unwrap_or("")
-                        .replace("\n", "<br>");
-                    output.push_str(&format!(
-                        "| `{}` | `{}` | {} |\n",
-                        field_name,
-                        format_type(field_type, data),
-                        docs
-                    ));
-                }
-            }
-        }
-    }
-
-    if union_.has_stripped_fields {
-        output.push_str("| *private fields* | ... | *Some fields have been omitted* |\n");
-    }
-
-    output.push('\n');
-
-    // Process impls
-    if !union_.impls.is_empty() {
-        output.push_str(&format!(
-            "{} Implementations\n\n",
-            "#".repeat(heading_level)
-        ));
-
-        // Group impls by trait
-        let mut trait_impls: std::collections::HashMap<String, Vec<Id>> =
-            std::collections::HashMap::new();
-        let mut inherent_impls: Vec<Id> = Vec::new();
-
-        for &impl_id in &union_.impls {
-            if let Some(impl_item) = data.index.get(&impl_id) {
-                if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                    if let Some(trait_) = &impl_.trait_ {
-                        let trait_name = trait_.path.clone();
-                        trait_impls.entry(trait_name).or_default().push(impl_id);
-                    } else {
-                        // Inherent impl
-                        inherent_impls.push(impl_id);
-                    }
-                }
-            }
-        }
-
-        // First list inherent impls
-        if !inherent_impls.is_empty() {
-            let methods_level = std::cmp::min(heading_level + 1, 6);
-            output.push_str(&format!("{} Methods\n\n", "#".repeat(methods_level)));
-            for &impl_id in &inherent_impls {
-                if let Some(impl_item) = data.index.get(&impl_id) {
-                    if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                        for &item_id in &impl_.items {
-                            if let Some(method_item) = data.index.get(&item_id) {
-                                if let ItemEnum::Function(_) = &method_item.inner {
-                                    if let Some(name) = &method_item.name {
-                                        output.push_str(&format!("- `{}`: ", name));
-                                        if let Some(docs) = &method_item.docs {
-                                            let first_line = docs.lines().next().unwrap_or("");
-                                            output.push_str(first_line);
-                                        }
-                                        output.push('\n');
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            output.push('\n');
-        }
-
-        // Then list trait impls
-        if !trait_impls.is_empty() {
-            let trait_impl_level = std::cmp::min(heading_level + 1, 6);
-            output.push_str(&format!(
-                "{} Trait Implementations\n\n",
-                "#".repeat(trait_impl_level)
-            ));
-            for (trait_name, impls) in trait_impls {
-                output.push_str(&format!("- **{}**\n", trait_name));
-                for &impl_id in &impls {
-                    if let Some(impl_item) = data.index.get(&impl_id) {
-                        if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                            for &item_id in &impl_.items {
-                                if let Some(method_item) = data.index.get(&item_id) {
-                                    if let Some(name) = &method_item.name {
-                                        output.push_str(&format!("  - `{}`: ", name));
-                                        if let Some(docs) = &method_item.docs {
-                                            let first_line = docs.lines().next().unwrap_or("");
-                                            output.push_str(first_line);
-                                        }
-                                        output.push('\n');
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            output.push('\n');
-        }
-    }
-}
-
-fn process_trait_details(output: &mut String, trait_: &Trait, data: &Crate, level: usize) {
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-    // Special traits info
-    if trait_.is_auto {
-        output.push_str("> This is an auto trait.\n\n");
-    }
-    if trait_.is_unsafe {
-        output.push_str("> This trait is unsafe to implement.\n\n");
-    }
-    if !trait_.is_dyn_compatible {
-        output.push_str(
-            "> This trait is not object-safe and cannot be used in dynamic trait objects.\n\n",
-        );
-    }
-
-    // Associated items
-    if !trait_.items.is_empty() {
-        // Group items by kind
-        let mut required_methods = Vec::new();
-        let mut provided_methods = Vec::new();
-        let mut assoc_types = Vec::new();
-        let mut assoc_consts = Vec::new();
-
-        for &item_id in &trait_.items {
-            if let Some(item) = data.index.get(&item_id) {
-                match &item.inner {
-                    ItemEnum::Function(function) => {
-                        if function.has_body {
-                            provided_methods.push(item_id);
-                        } else {
-                            required_methods.push(item_id);
-                        }
-                    }
-                    ItemEnum::AssocType { .. } => assoc_types.push(item_id),
-                    ItemEnum::AssocConst { value, .. } => {
-                        if value.is_some() {
-                            // Has a default value
-                            provided_methods.push(item_id);
-                        } else {
-                            assoc_consts.push(item_id);
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        // Required items
-        if !required_methods.is_empty() || !assoc_types.is_empty() || !assoc_consts.is_empty() {
-            output.push_str(&format!("{} Required Items\n\n", "#".repeat(heading_level)));
-
-            if !assoc_types.is_empty() {
-                output.push_str(&format!(
-                    "{} Associated Types\n\n",
-                    "#".repeat(heading_level + 1)
-                ));
-                for &type_id in &assoc_types {
-                    if let Some(type_item) = data.index.get(&type_id) {
-                        if let Some(name) = &type_item.name {
-                            output.push_str(&format!("- `{}`", name));
-                            if let Some(docs) = &type_item.docs {
-                                if let Some(first_line) = docs.lines().next() {
-                                    if !first_line.trim().is_empty() {
-                                        output.push_str(&format!(": {}", first_line));
-                                    }
-                                }
-                            }
-                            output.push('\n');
-                        }
-                    }
-                }
-                output.push('\n');
-            }
-
-            if !assoc_consts.is_empty() {
-                output.push_str(&format!(
-                    "{} Associated Constants\n\n",
-                    "#".repeat(heading_level + 1)
-                ));
-                for &const_id in &assoc_consts {
-                    if let Some(const_item) = data.index.get(&const_id) {
-                        if let Some(name) = &const_item.name {
-                            output.push_str(&format!("- `{}`", name));
-                            if let Some(docs) = &const_item.docs {
-                                if let Some(first_line) = docs.lines().next() {
-                                    if !first_line.trim().is_empty() {
-                                        output.push_str(&format!(": {}", first_line));
-                                    }
-                                }
-                            }
-                            output.push('\n');
-                        }
-                    }
-                }
-                output.push('\n');
-            }
-
-            if !required_methods.is_empty() {
-                output.push_str(&format!(
-                    "{} Required Methods\n\n",
-                    "#".repeat(heading_level + 1)
-                ));
-                for &method_id in &required_methods {
-                    if let Some(method_item) = data.index.get(&method_id) {
-                        if let Some(name) = &method_item.name {
-                            output.push_str(&format!("- `{}`", name));
-                            if let Some(docs) = &method_item.docs {
-                                if let Some(first_line) = docs.lines().next() {
-                                    if !first_line.trim().is_empty() {
-                                        output.push_str(&format!(": {}", first_line));
-                                    }
-                                }
-                            }
-                            output.push('\n');
-                        }
-                    }
-                }
-                output.push('\n');
-            }
-        }
-
-        // Provided items
-        if !provided_methods.is_empty() {
-            output.push_str(&format!(
-                "{} Provided Methods\n\n",
-                "#".repeat(heading_level)
-            ));
-            for &method_id in &provided_methods {
-                if let Some(method_item) = data.index.get(&method_id) {
-                    if let ItemEnum::Function(_) = &method_item.inner {
-                        // Format method signature
-                        let mut method_signature = String::new();
-                        format_item_signature(&mut method_signature, method_item, data);
-
-                        // Output with proper code block formatting
-                        output.push_str("- ```rust\n  ");
-                        output.push_str(&method_signature.trim());
-                        output.push_str("\n  ```");
-
-                        // Add documentation if available
-                        if let Some(docs) = &method_item.docs {
-                            if let Some(first_line) = docs.lines().next() {
-                                if !first_line.trim().is_empty() {
-                                    output.push_str(&format!("\n  {}", first_line));
-                                }
-                            }
-                        }
-                        output.push_str("\n\n");
-                    }
-                }
-            }
-        }
-    }
-
-    // Implementations
-    if !trait_.implementations.is_empty() {
-        output.push_str(&format!(
-            "{} Implementations\n\n",
-            "#".repeat(heading_level)
-        ));
-        output.push_str("This trait is implemented for the following types:\n\n");
-
-        for &impl_id in &trait_.implementations {
-            if let Some(impl_item) = data.index.get(&impl_id) {
-                if let ItemEnum::Impl(impl_) = &impl_item.inner {
-                    output.push_str(&format!("- `{}`", format_type(&impl_.for_, data)));
-                    // Add generics if present
-                    if !impl_.generics.params.is_empty() {
-                        let mut generics_str = String::new();
-                        format_generics(&mut generics_str, &impl_.generics, data);
-                        if generics_str != "<>" {
-                            output.push_str(" with ");
-                            output.push_str(&generics_str);
-                        }
-                    }
-                    output.push('\n');
-                }
-            }
-        }
-        output.push('\n');
-    }
-}
-
-fn process_impl_details(output: &mut String, impl_: &Impl, data: &Crate, level: usize) {
-    // Cap heading level at 6 (maximum valid Markdown heading level)
-    let heading_level = std::cmp::min(level, 6);
-    // List all items in the impl
-    if !impl_.items.is_empty() {
-        output.push_str(&format!(
-            "{} Associated Items\n\n",
-            "#".repeat(heading_level)
-        ));
-
-        // Group by kind
-        let mut methods = Vec::new();
-        let mut assoc_types = Vec::new();
-        let mut assoc_consts = Vec::new();
-
-        for &item_id in &impl_.items {
-            if let Some(item) = data.index.get(&item_id) {
-                match &item.inner {
-                    ItemEnum::Function(_) => methods.push(item_id),
-                    ItemEnum::AssocType { .. } => assoc_types.push(item_id),
-                    ItemEnum::AssocConst { .. } => assoc_consts.push(item_id),
-                    _ => {}
-                }
-            }
-        }
-
-        if !assoc_types.is_empty() {
-            output.push_str(&format!(
-                "{} Associated Types\n\n",
-                "#".repeat(heading_level + 1)
-            ));
-            for &type_id in &assoc_types {
-                if let Some(assoc_item_ref) = data.index.get(&type_id) {
-                    let resolved_assoc_info = ResolvedItemInfo {
-                        original_item: assoc_item_ref,
-                        effective_name: assoc_item_ref.name.clone(),
-                        reexport_source_canonical_path: None,
-                    };
-                    process_item(output, &resolved_assoc_info, data, level + 1);
-                }
-            }
-        }
-
-        if !assoc_consts.is_empty() {
-            output.push_str(&format!(
-                "{} Associated Constants\n\n",
-                "#".repeat(heading_level + 1)
-            ));
-            for &const_id in &assoc_consts {
-                if let Some(assoc_item_ref) = data.index.get(&const_id) {
-                    let resolved_assoc_info = ResolvedItemInfo {
-                        original_item: assoc_item_ref,
-                        effective_name: assoc_item_ref.name.clone(),
-                        reexport_source_canonical_path: None,
-                    };
-                    process_item(output, &resolved_assoc_info, data, level + 1);
-                }
-            }
-        }
-
-        if !methods.is_empty() {
-            output.push_str(&format!("{} Methods\n\n", "#".repeat(heading_level + 1)));
-            for &method_id in &methods {
-                if let Some(assoc_item_ref) = data.index.get(&method_id) {
-                    let resolved_assoc_info = ResolvedItemInfo {
-                        original_item: assoc_item_ref,
-                        effective_name: assoc_item_ref.name.clone(),
-                        reexport_source_canonical_path: None,
-                    };
-                    process_item(output, &resolved_assoc_info, data, level + 1);
-                }
-            }
-        }
-    }
-
-    // If this is a trait impl, list the provided trait methods that aren't overridden
-    if impl_.trait_.is_some() && !impl_.provided_trait_methods.is_empty() {
-        output.push_str(&format!(
-            "{} Provided Trait Methods\n\n",
-            "#".repeat(heading_level)
-        ));
-        output.push_str("The following methods are available through the trait but not explicitly implemented:\n\n");
-
-        for provided_method in &impl_.provided_trait_methods {
-            output.push_str(&format!("- `{}`\n", provided_method));
-        }
-
-        output.push('\n');
-    }
-
-    // If this is a blanket impl, mention it
-    if let Some(blanket_type) = &impl_.blanket_impl {
-        output.push_str(&format!(
-            "This is a blanket implementation for all types that match: `{}`\n\n",
-            format_type(blanket_type, data)
-        ));
-    }
 }
